+++ changelog.md
@@ -7,10 +7,15 @@ The changelog follows a few basic guidelines
 - If the creator has not previously existed, the words "Added content by XXX," where XXX is the creator's name are used. The given entry follows the previous constraint
 - If the creator has previously existed but the newly-created folder has not, the words "Added YYY," where YYY is the new folder name are used. The given entry follows the ordering constraint described above
 - If the creator's content has been moved or merged with another author, the words "Moved XXX to YYY," where XXX is the original path and YYY is the new path are used. The given entry follows the ordering constraint described above
-## 07-5-2021
+## 16-09-2021
+- Replaced Garrett Brown's old RS Games recording with a new one.
+- Added Garrett Brown's Twitch
+- Added Holy Christ to Everstone's author page.
+- Added website creators to a footer
+## 07-09-2021
 ### Key is Full
 - Added BK2 Failures
-## 06-05-2021
+## 06-09-2021
 - Added content by Garrett Brown
 ### Key is full
 - Code Green Fifty Archives/2017-6-2 Code Green Gets Randomly Emo while Looking for Crystals